refactor(eve): extract dev URL target parsing
